+++ fuse_tutorials/src/beacon_publisher.cpp
@@ -31,7 +31,7 @@
  *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  *  POSSIBILITY OF SUCH DAMAGE.
  */
-#include <fuse_tutorials/beacon_publisher.h>
+#include <fuse_tutorials/beacon_publisher.hpp>
 
 #include <fuse_core/async_publisher.hpp>
 #include <fuse_core/graph.hpp>
